# [New feature][Performance] Introduce a caching mechanism to ListView to reduce unnecessary layout during scrolling