@xu-baolin 2023-10-12
This is my thoughts.
This feature is enabled optional by the developer and only applies if the developer is sure that the children's sizes will not changed during the life cycle. I think there are many such scenarios, for example, on the web, once the child renders, the size will not change unless the user refreshes the page. In these scenarios, caching can greatly improve scrolling performance.






________________________________
来自TT移动邮箱 / From TT mailbox





________________________________
发件人: chunhtai ***@***.***>
发送时间: 2023年10月13日 03:53
收件人: flutter/flutter ***@***.***>
抄送: 徐宝林(Ava) ***@***.***>; Author ***@***.***>
主题: Re: [flutter/flutter] [New feature][Performance] Introduce a caching mechanism to ListView to reduce unnecessary layout during scrolling (PR #134864)


@chunhtai commented on this pull request.

This api seems to be a bit dangerous. It is hard to predict child size change. I am not sure if this is something we want to pursue. There is also a PreferredSizeWidget, but that will require developer to give the size beforehand.

________________________________

In packages/flutter/lib/src/rendering/sliver_list.dart<https://github.com/flutter/flutter/pull/134864#discussion_r1357281415>:

> @@ -38,24 +47,171 @@ class RenderSliverList extends RenderSliverMultiBoxAdaptor {
   /// the main axis.
   RenderSliverList({
     required super.childManager,
-  });
+    bool enableItemExtentsCaching = false,
+  }) : _enableItemExtentsCaching = enableItemExtentsCaching;
+
+  /// Whether to cache the main axis extents of all laid out children.
+  ///
+  /// If true, the main axis extent of child will be cached as
+  /// index/extent pairs when the child first laid out.
+  ///
+  /// If false, the main axis extents of children will not be cached and
+  /// previously cached data will be cleared.
+  ///
+  /// The cached data will be used to efficiently determine the target children
+  /// within the lazy loading scope, which helps improve layout performance.


I am not sure how this would work, even if the SliverChildListDelegate doesn't change, the child itself can still change size.

________________________________

In packages/flutter/lib/src/rendering/sliver_list.dart<https://github.com/flutter/flutter/pull/134864#discussion_r1357292493>:

> +  }
+
+  void _performLoadingScopeChildrenLayoutIfNeeded() {
+    if (!enableItemExtentsCaching) {
+      return;
+    }
+    for (RenderBox? child = firstChild; child != null ; child = childAfter(child)) {
+      final int index = indexOf(child);
+      if (_childrenWithoutLayout.contains(index)) {
+        child.layout(_childConstraints, parentUsesSize: true);
+        // Throw if the real extent differs from the cached value.
+        assert(() {
+          if (super.paintExtentOf(child!) != _cachedItemExtents[index]!.mainAxisExtent) {
+            throw FlutterError(
+              'The item(index: $index) main axis extent has changed after caching, if this is '
+              'expected, change its Key to update the cached value.',


This seems dangerous if developers are not sure what they are doing. It is also a bit confusing that it would require manual invalidation of caches. I think what would happen is developer will just use the invalidating caches everywhere so that it doesn't crash.

―
Reply to this email directly, view it on GitHub<https://github.com/flutter/flutter/pull/134864#pullrequestreview-1674971562>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/AOR66GDACE653AFWYO4JCMLX7BDFJANCNFSM6AAAAAA42VIBIE>.
You are receiving this because you authored the thread.Message ID: ***@***.***>

________________________________
OPPO

本电子邮件及其附件含有OPPO公司的保密信息，仅限于邮件指明的收件人（包含个人及群组）使用。禁止任何人在未经授权的情况下以任何形式使用。如果您错收了本邮件，切勿传播、分发、复制、印刷或使用本邮件之任何部分或其所载之任何内容，并请立即以电子邮件通知发件人并删除本邮件及其附件。
网络通讯固有缺陷可能导致邮件被截留、修改、丢失、破坏或包含计算机病毒等不安全情况，OPPO对此类错误或遗漏而引致之任何损失概不承担责任并保留与本邮件相关之一切权利。
除非明确说明，本邮件及其附件无意作为在任何国家或地区之要约、招揽或承诺，亦无意作为任何交易或合同之正式确认。 发件人、其所属机构或所属机构之关联机构或任何上述机构之股东、董事、高级管理人员、员工或其他任何人（以下称“发件人”或“OPPO”）不因本邮件之误送而放弃其所享之任何权利，亦不对因故意或过失使用该等信息而引发或可能引发的损失承担任何责任。
文化差异披露：因全球文化差异影响，单纯以YES\OK或其他简单词汇的回复并不构成发件人对任何交易或合同之正式确认或接受，请与发件人再次确认以获得明确书面意见。发件人不对任何受文化差异影响而导致故意或错误使用该等信息所造成的任何直接或间接损害承担责任。
This e-mail and its attachments contain confidential information from OPPO, which is intended only for the person or entity whose address is listed above. Any use of the information contained herein in any way (including, but not limited to, total or partial disclosure, reproduction, or dissemination) by persons other than the intended recipient(s) is prohibited. If you are not the intended recipient, please do not read, copy, distribute, or use this information. If you have received this transmission in error, please notify the sender immediately by reply e-mail and then delete this message.
Electronic communications may contain computer viruses or other defects inherently, may not be accurately and/or timely transmitted to other systems, or may be intercepted, modified ,delayed, deleted or interfered. OPPO shall not be liable for any damages that arise or may arise from such matter and reserves all rights in connection with the email.
Unless expressly stated, this e-mail and its attachments are provided without any warranty, acceptance or promise of any kind in any country or region, nor constitute a formal confirmation or acceptance of any transaction or contract. The sender, together with its affiliates or any shareholder, director, officer, employee or any other person of any such institution (hereinafter referred to as "sender" or "OPPO") does not waive any rights and shall not be liable for any damages that arise or may arise from the intentional or negligent use of such information.
Cultural Differences Disclosure: Due to global cultural differences, any reply with only YES\OK or other simple words does not constitute any confirmation or acceptance of any transaction or contract, please confirm with the sender again to ensure clear opinion in written form. The sender shall not be responsible for any direct or indirect damages resulting from the intentional or misuse of such information.