@chunhtai 2023-10-12
I can see how it can be useful in some circumstances, but I am still not convince we should provide it in the framework as it can be easily misused. Have you considered put this into a third party package instead?